# backend/openapi.yaml
openapi: 3.0.1
info:
  title: AguaVigia CTG — API publica
  description: |-
    Monitoreo ciudadano del acueducto de Cartagena de Indias.

    Plataforma ciudadana e independiente: no esta afiliada a Aguas de Cartagena
    S.A. E.S.P. ni a ninguna entidad distrital.
  contact:
    name: AguaVigia CTG
  license:
    name: Proyecto de aula
  version: 1.0.0
servers:
- url: /
  description: Mismo origen que sirve la API
...
paths:
...
  /api/veedor/cortes/{id}:
    get:
      tags:
      - Veedor - Cortes
      summary: Consultar un corte por su identificador
      operationId: consultar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Corte encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CorteRespuesta"
        "404":
          description: No existe un corte con ese id
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "Sin sesión, token inválido, caducado o revocado"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: La sesión no tiene el permiso que exige esta operación
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
    CorteRespuesta:
      type: object
      properties:
        id:
          type: string
        sectoresAfectados:
          type: array
          items:
            type: string
        inicio:
          type: string
          format: date-time
        finPrometido:
          type: string
          format: date-time
        finReal:
          type: string
          description: Nulo mientras el corte sigue abierto
          format: date-time
        causa:
          type: string
        origen:
          type: string
        estado:
          type: string
      description: Corte oficial (RF016-RF017)
...
  securitySchemes:
    bearerAuth:
      type: http
      description: |-
        Token que devuelve POST /api/veedor/sesion, en la cabecera
        `Authorization: Bearer <token>`. Caduca a las 8 horas y no se renueva: al
        caducar hay que volver a iniciar sesión.
      scheme: bearer
      bearerFormat: JWT